Change minister models
