+++ app/models/mandate.js
@@ -1,7 +1,7 @@
-import Model, { hasMany, belongsTo } from '@ember-data/model';
+import { hasMany, belongsTo } from '@ember-data/model';
+import PostModel from './post';
 
-export default class MandateModel extends Model {
-  @belongsTo('board-position') roleBoard;
+export default class MandateModel extends PostModel {
   @belongsTo('governing-body', { inverse: 'mandates' }) governingBody;
   @hasMany('mandatory', { inverse: 'mandate' }) heldBy;
 }
