+++ app/models/minister-position-functions.js
@@ -1,0 +1,5 @@
+import Model, { attr } from '@ember-data/model';
+
+export default class MinisterPositionFunctionsModel extends Model {
+  @attr label;
+}
