+++ app/models/minister-position.js
@@ -1,5 +1,11 @@
-import Model, { attr } from '@ember-data/model';
+import { belongsTo, hasMany } from '@ember-data/model';
+import PostModel from './post';
 
-export default class MinisterPositionModel extends Model {
-  @attr label;
+export default class MinisterPositionModel extends PostModel {
+  @belongsTo('minister-position-functions') function;
+  @hasMany('minister', { inverse: 'ministerPosition' }) heldyByMinisters;
+  @belongsTo('worship-service', { inverse: 'ministerPositions' })
+  worshipService;
+  @belongsTo('representative-body', { inverse: 'ministerPositions' })
+  representativeBody;
 }
