+++ app/models/representative-body.js
@@ -1,6 +1,7 @@
-import { belongsTo } from '@ember-data/model';
+import { belongsTo, hasMany } from '@ember-data/model';
 import OrganizationModel from './organization';
 
 export default class RepresentativeBodyModel extends OrganizationModel {
   @belongsTo('recognized-worship-type') recognizedWorshipType;
+  @hasMany('minister-positions') ministerPositions;
 }
