+++ app/routes/administrative-units/administrative-unit/ministers.js
@@ -1,24 +1,3 @@
 import Route from '@ember/routing/route';
-import { inject as service } from '@ember/service';
 
-export default class AdministrativeUnitsAdministrativeUnitMinistersRoute extends Route {
-  @service store;
-
-  async model() {
-    let { id: administrativeUnitId } = this.paramsFor(
-      'administrative-units.administrative-unit'
-    );
-
-    let administrativeUnit = await this.store.findRecord(
-      'administrative-unit',
-      administrativeUnitId,
-      {
-        reload: true,
-      }
-    );
-
-    return {
-      administrativeUnit: administrativeUnit,
-    };
-  }
-}
+export default class AdministrativeUnitsAdministrativeUnitMinistersRoute extends Route {}
