+++ app/routes/administrative-units/administrative-unit/ministers/index.js
@@ -1,3 +1,22 @@
 import Route from '@ember/routing/route';
+import { inject as service } from '@ember/service';
 
-export default class AdministrativeUnitsAdministrativeUnitMinistersIndexRoute extends Route {}
+export default class AdministrativeUnitsAdministrativeUnitMinistersIndexRoute extends Route {
+  @service store;
+
+  async model() {
+    let { id: administrativeUnitId } = this.paramsFor(
+      'administrative-units.administrative-unit'
+    );
+
+    let administrativeUnit = await this.store.findRecord(
+      'worship-service',
+      administrativeUnitId
+    );
+
+    return {
+      administrativeUnit: administrativeUnit,
+      ministerPositions: await administrativeUnit.ministerPositions,
+    };
+  }
+}
