+++ app/templates/administrative-units/administrative-unit.hbs
@@ -41,7 +41,9 @@
           </AuNavigationLink>
         </li>
         <li class="au-c-list-navigation__item">
-          <AuNavigationLink>
+          <AuNavigationLink
+            @route="administrative-units.administrative-unit.ministers"
+          >
             Bedienaren
           </AuNavigationLink>
         </li>
