+++ app/templates/administrative-units/administrative-unit/ministers.hbs
@@ -3,6 +3,7 @@
 {{breadcrumb
   "Bedienaren"
   route="administrative-units.administrative-unit.ministers"
+  model=(params-for "administrative-units.administrative-unit" param="id")
 }}
 
-{{outlet}}  
+{{outlet}}
