+++ app/templates/administrative-units/administrative-unit/ministers/index.hbs
@@ -6,13 +6,19 @@
           Bedienaren
         </AuHeading>
       </AuToolbarGroup>
+      <AuToolbarGroup>
+        {{! TODO: Replace this with a link once the "new" route is created }}
+        <AuButton @icon="add" @iconAlignment="left">
+          Nieuw
+        </AuButton>
+      </AuToolbarGroup>
     </AuToolbar>
     <AuHeading @level="2" @skin="3" class="au-u-margin-top-tiny">
       {{@model.administrativeUnit.name}}
     </AuHeading>
   </div>
 
-  <AuDataTable @noDataMessage="Geen bestuursorganen" as |t|>
+  <AuDataTable @content={{this.ministerPositions}} @noDataMessage="Geen bedienaren" as |t|>
     <t.content as |c|>
       <c.header>
         <th class="au-c-data-table__header-title">Voornaam</th>
@@ -21,13 +27,31 @@
         <th class="au-c-data-table__header-title">Status</th>
         <th></th>
       </c.header>
-      <c.body>
-        <td><AuLink @route="administrative-units.administrative-unit.ministers">Voornaam</AuLink></td>
-        <td><AuLink @route="administrative-units.administrative-unit.ministers">Achternaam</AuLink></td>
-        <td><AuLink @route="administrative-units.administrative-unit.ministers">Rol</AuLink></td>
-        <td><AuPill @skin={{"success"}}>Actief</AuPill></td>
-        <td><AuLink @route="administrative-units.administrative-unit.ministers">Bewerk</AuLink></td>
+      <c.body as |ministerPosition|>
+        {{#let ministerPosition.heldyByMinisters.firstObject as |minister|}}
+          <td>
+            <AuLink @route="administrative-units.administrative-unit.ministers">
+              {{minister.person.givenName}}
+            </AuLink>
+          </td>
+          <td>
+            <AuLink @route="administrative-units.administrative-unit.ministers">
+              {{minister.person.familyName}}
+            </AuLink>
+          </td>
+          <td>
+            <AuLink @route="administrative-units.administrative-unit.ministers">
+              {{ministerPosition.function.label}}
+            </AuLink>
+          </td>
+          <td><AuPill @skin={{"success"}}>Actief</AuPill></td>
+          <td>
+            <AuLink @route="administrative-units.administrative-unit.ministers">
+              Bewerk
+            </AuLink>
+          </td>
+        {{/let}}
       </c.body>
     </t.content>
   </AuDataTable>
-</div> 
+</div>
